// src/types.ts
export interface UserOperation {
  sender: string;
  nonce: string;
  initCode: string;
  callData: string;
  callGasLimit: string;
  verificationGasLimit: string;
  preVerificationGas: string;
  maxFeePerGas: string;
  maxPriorityFeePerGas: string;
  paymasterAndData: string;
  signature: string;
}

export interface Config {
  rpcUrl: string;
  privateKey: string;
  bundlerUrl: string;
  factoryAddress: string;
  entryPointAddress: string;
}